Big change to support keypad and stand-alone mode
This change incorporates code from
http://hackaday.com/2012/07/27/bending-a-home-security-control-panel-to-your-will/
to interface to a DSC PC1500RK keypad.

The keypad allows the node to function in standalone mode.  I've added
logic to make it function similarly to the original security system but
with interaction via the JeeNode's radio.  Messages are sent when an
alarm is triggered or any time a sensor value changes.

In addition, the node now operates in an event-based manner.  Instead of
periodically sending sensor update packets, the node now sends sensor
updates only when they change.  To help with this new model there is now
a "sensor update" command that will force a sensor update packet to be
sent.

The LabVIEW interface app has been updated to work with the new
firmware.

There are still features that should be added, most notably more
security on the RF side.